requirements.txt
pandas
torch
patsy
statsmodels
matplotlib
pyyaml